Extracted some reusable components to common directory
