+++ apps/uriparser-test.cpp
@@ -1,1 +1,1 @@
-uriparser.cpp
+../common/uriparser.cpp
