+++ common/appcommon.hpp
